ENG-382: fix(CAdapter): fix cToken interface

+++ pkg/core/src/adapters/compound/CAdapter.sol
@@ -20,7 +20,7 @@ interface CTokenInterface {
     /// @dev returns the current exchange rate as an uint, scaled by 1 * 10^(18 - 8 + Underlying Token Decimals).
     function exchangeRateCurrent() external returns (uint256);
 
-    function decimals() external returns (uint256);
+    function decimals() external returns (uint8);
 
     function underlying() external view returns (address);
